added user authentication

+++ app/Http/Controllers/ProjectsController.php
@@ -9,14 +9,21 @@ use Illuminate\Http\Request;
 
 class ProjectsController extends Controller
 {
-    public function index()
+
+    public function __construct()
     {
-      $projects = Project::all();
+    //  $this->middleware('can:update,project');
+    }
+
+    public function index(Project $project)
+    {
+
+      $projects = Project::where('owner_id', auth()->id())->get();
 
       return view('projects.index', compact('projects'));
     }
 
-    public function create()
+    public function create(Project $project)
     {
       return view('projects.create');
     }
@@ -24,36 +31,53 @@ class ProjectsController extends Controller
     public function store(Project $project)
     {
 
-      Project::create(
-        request()->validate([
-          'title' => ['required', 'min:3', 'max:128'],
-          'description' => ['required', 'min:3', 'max:255']
-        ])
-      );
+      $attributes = request()->validate([
+        'title' => ['required', 'min:3', 'max:128'],
+        'description' => ['required', 'min:3', 'max:255'],
+      ]);
+
+      $attributes['owner_id'] = auth()->id();
+
+      Project::create($attributes);
 
       return redirect('/projects');
     }
 
     public function show(Project $project)
     {
+      $this->authorize('update', $project);
+
       return view('projects.show', compact('project'));
     }
 
     public function edit(Project $project)
     {
+      $this->authorize('update', $project);
+
       return view('projects.edit', compact('project'));
     }
 
     public function update(Project $project)
     {
+      $this->authorize('update', $project);
 
-      $project->update(request(['title', 'description']));
+      $attributes = request()->validate([
+        'title' => ['required', 'min:3', 'max:128'],
+        'description' => ['required', 'min:3', 'max:255'],
+      ]);
+
+      $attributes['owner_id'] = auth()->id();
+
+      $project->update($attributes);
 
       return redirect('/projects');
     }
 
     public function destroy(Project $project)
     {
+
+      $this->authorize('update', $project);
+
       $project->delete();
 
       return redirect('/projects');
